moved module to assemblyai directory and modified transcribe.py to use functions defined in module
